Handling cases of mismatched data types

+++ src/construct_cells.cpp
@@ -13,14 +13,14 @@ void getTopCells(Simulation *sim, Grid *grid) {
   // Unpack the simulation information we need
   std::vector<Cell>& cells = sim->cells;
   const double width[3] = {sim->width[0], sim->width[1], sim->width[2]};
-  const int nr_cells = sim->nr_cells;
+  const size_t nr_cells = sim->nr_cells;
   const int cdim[3] = {sim->cdim[0], sim->cdim[1], sim->cdim[2]};
   const std::vector<int> counts = sim->cell_part_counts;
 
 // Loop over the cells and create them, storing the counts for domain
 // decomposition
 #pragma omp parallel for
-  for (int cid = 0; cid < nr_cells; cid++) {
+  for (size_t cid = 0; cid < nr_cells; cid++) {
 
     // Get integer coordinates of the cell
     int i = cid / (cdim[1] * cdim[2]);
@@ -64,7 +64,7 @@ void getTopCells(Simulation *sim, Grid *grid) {
   // Loop over the cells attaching the pointers the neighbouring cells (taking
   // into account periodic boundary conditions)
 #pragma omp parallel for
-  for (int cid = 0; cid < nr_cells; cid++) {
+  for (size_t cid = 0; cid < nr_cells; cid++) {
 
     // Get integer coordinates of the cell
     int i = cid / (cdim[1] * cdim[2]);
@@ -111,12 +111,12 @@ void splitCells(Simulation *sim) {
   Metadata *metadata = &Metadata::getInstance();
 
   // Unpack the cells
-  const int nr_cells = sim->nr_cells;
+  const size_t nr_cells = sim->nr_cells;
   std::vector<Cell>& cells = sim->cells;
 
   // Loop over the cells and split them
 #pragma omp parallel for
-  for (int cid = 0; cid < nr_cells; cid++) {
+  for (size_t cid = 0; cid < nr_cells; cid++) {
 
 #ifdef WITH_MPI
     // Skip cells that aren't on this rank and aren't proxies
